Update flow session handling
Setting the entire session as a cookie was using up too much space and
causing 502s on the QA site.  Update to only store the current
experience.  Update the middleware to create the session in flow so a
users first page hit in foreign countries doesn't show the base
currency.

no changelog

FLOW-41