Remove Extra Configuration Setting
This was a previous Workarea config that prevented refund tests from running
when they didn't need to be, but those tests have been removed from
Workarea base, so the configuration is no longer necessary here. Since
`FlowPaymentIntegrationTest` re-implements those tests, we don't need
the setting or the tests to be pasted in here.

FLOW-1