Update Webhooks Setup
The webhook shared secret is now generated by the Workarea application
automatically as part of the install procedure for the plugin, right
before webhooks are created on Flow's side. Update README to include
information on new configuration settings, remove notes on the webhook
secret, and reformat the docs so they're friendly to total newbies as
well as an old Linuxbeard. Fix some typos and grammar in the README as
well. This also adds the `LICENSE` file to the repository for future open
source goodness.

FLOW-2