Add FlowJS for Real-Time Pricing Updates
It's possible that prices could change between the time they are synced
over to us via the (upcoming) CSV FTP upload, and they could change for
experiences that aren't explicitly configured in the Flow Console. In
order to address this, the FlowJS library has been integrated to provide
pricing updates to each product as it appears on screen. A future code
change will (probably) be made that addresses the consumption of this CSV
file, which is how we'll be obtaining prices from Flow instead of using
the webhooks.

In addition, this removes the old "Flow Beacon" and "Flow Country
Picker" scripts, since `flow.js` handles both of those use cases
out-of-box.

This JS integration is fully opt-in, and won't activate unless you
configure `config.flow_io.enable_javascript = true`.

FLOW-4