Revise Cart Pricing
Workarea now sends over the prices that were actually displayed to the
user in the cart to Flow's backend, eliminating any kind of price
discrepency between Flow and the Workarea application. Additionally, the
current experience is now being used to dictate the currency of all item
prices and discounts sent over in the `OrderPutForm`. In order to
prevent unnecessary test failures related to the year 2020, this gem now
depends on v3.3.36 or higher of the platform.

FLOW-9