Fix Base Item Price In Line Items
Passing `cents` into the price amount for a line item caused prices on
Flow's side to be way higher than they should be. This should have been
a floating point value equal to the price of the item. It has been
changed as such and should fix the issues experienced in QA for checking
out Flow items.

FLOW-11